feat: rename NGC compiler plugin
