+++ packages/angular/vite/src/index.ts
@@ -1,3 +1,3 @@
 export * from './lib/vite';
-export * from './lib/angular-vite-compiler';
+export * from './lib/angular-ngc-vite-compiler';
 export * from './lib/vite-plugin-angular';
