+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/angular-vite-plugin.ts
