+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/cache.ts
