+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/host.ts
