+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/index.ts
