+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/ngcc_processor.ts
