+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/path.ts
