+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/plugin-options.ts
