+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/readme.md
@@ -1,0 +1,5 @@
+# Angular NGC vite plugin
+
+This plugin was develop for prototyping. This is no longer maintained. If you want a NGC full application complier that works in Vite then feel free to still use this version. Feel free to create PR to fix any issues.
+
+The future was deemed away from full application compiling to a file by file a pure Vite/SWC/ESBuild way of working
