+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/transformers/index.ts
