+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/transformers/remove-ivy-jit-support-calls.ts
