+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/transformers/replace_resources.ts
