+++ packages/angular/vite/src/lib/angular-ngc-vite-compiler/utils/load-tsconfig.ts
